feat(common): change the request id as Enum
Core lightning change the id typ from int to string,
to allow a better identification of quest from plugin.

So this commit support this new kind of id and remove the
str type inside the request in favor of String.

The main motivation to switch from str to String is to remove
the lifetime that did not work so well with serde.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>