feat(proc_macro): Minimal and complete conversion to kproc-macros
This commit deserves a bit more attention in terms of its description.

I've never been a fan of `syn` because it's a large dependency and
too complex for my use case. Additionally, I cannot remove `proc_macro2`
if I want to. For this reason, I chose to write
`kproc-macros` (https://github.com/vincenzopalazzo/kproc-macros), which
implements a very lightweight parser for Rust procedural macros and
provides an API to extend the procedural macros API to
build your own parser.

This commit introduces a basic version of the `kproc-macros`
crate that allows me to perform debugging on it.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>